updating readme for new design
Previous design had issue. and few of the classes were holding refrences of other classes, which incressed the coupling of the design. and also breaking SRP (single responsible princible). Most of the issue has been resolved in the new design. Classes are made as independent as possible with defined responsiblity.